feat: add OrderBy trait
